% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmspcov.R
\name{bmspcov}
\alias{bmspcov}
\title{Bayesian Sparse Covariance Estimation}
\usage{
bmspcov(X, Sigma, prior = list(), nsample = list())
}
\arguments{
\item{X}{a n \eqn{\times} p data matrix with column mean zero.}

\item{Sigma}{an initial guess for Sigma.}

\item{prior}{a list giving the prior information.
The list includes the following parameters (with default values in parentheses):
\code{a (1/2)} and \code{b (1/2)} giving the shape parameters for beta distribution,
\code{lambda (1)} giving the hyperparameter for the diagonal elements,
\code{tau1sq (10000/(n*p^4))} giving the hyperparameter for the shrinkage prior of covariance.}

\item{nsample}{a list giving the MCMC parameters.
The list includes the following integers (with default values in parentheses):
\code{burnin (1000)} giving the number of MCMC samples in transition period,
\code{nmc (1000)} giving the number of MCMC samples for analysis.}
}
\value{
\item{Sigma}{a nmc \eqn{\times} p(p+1)/2 matrix including lower triangular elements of covariance matrix.}
\item{Phi}{a nmc \eqn{\times} p(p+1)/2 matrix including shrinkage parameters corresponding to lower triangular elements of covariance matrix.}
\item{p}{dimension of covariance matrix.}
}
\description{
Provides a Bayesian sparse and positive definite estimate of a covariance matrix via the beta-mixture shrinkage prior.
}
\details{
Lee, Jo and Lee (2022) proposed the beta-mixture shrinkage prior for estimating a sparse and positive definite covariance matrix.
The beta-mixture shrinkage prior for \eqn{\Sigma = (\sigma_{jk})} is defined as
\deqn{
 \pi(\Sigma) = \frac{\pi^{u}(\Sigma)I(\Sigma \in C_p)}{\pi^{u}(\Sigma \in C_p)}, ~ C_p = \{\mbox{ all } p \times p \mbox{ positive definite matrices }\},
}
where \eqn{\pi^{u}(\cdot)} is the unconstrained prior given by
\deqn{
\pi^{u}(\sigma_{jk} \mid \rho_{jk}) = N\left(\sigma_{jk} \mid 0, \frac{\rho_{jk}}{1 - \rho_{jk}}\tau_1^2\right)}
\deqn{
\pi^{u}(\rho_{jk}) = Beta(\rho_{jk} \mid a, b), ~ \rho_{jk} = 1 - 1/(1 + \phi_{jk})}
\deqn{
\pi^{u}(\sigma_{jj}) = Exp(\sigma_{jj} \mid \lambda).
}
For more details, see Lee, Jo and Lee (2022).
}
\examples{

set.seed(1)
n <- 20
p <- 5

# generate a sparse covariance matrix:
True.Sigma <- matrix(0, nrow = p, ncol = p)
diag(True.Sigma) <- 1
Values <- -runif(n = p*(p-1)/2, min = 0.2, max = 0.8)
nonzeroIND <- which(rbinom(n=p*(p-1)/2,1,prob=1/p)==1)
zeroIND = (1:(p*(p-1)/2))[-nonzeroIND]
Values[zeroIND] <- 0
True.Sigma[lower.tri(True.Sigma)] <- Values
True.Sigma[upper.tri(True.Sigma)] <- t(True.Sigma)[upper.tri(True.Sigma)]
if(min(eigen(True.Sigma)$values) <= 0){
  delta <- -min(eigen(True.Sigma)$values) + 1.0e-5
  True.Sigma <- True.Sigma + delta*diag(p)
}

# generate a data
X <- MASS::mvrnorm(n = n, mu = rep(0, p), Sigma = True.Sigma)

# compute sparse, positive covariance estimator:
fout <- bspcov::bmspcov(X = X, Sigma = diag(diag(cov(X))))
post.est.m <- bspcov::estimate(fout)
sqrt(mean((post.est.m - True.Sigma)^2))
sqrt(mean((cov(X) - True.Sigma)^2))

}
\references{
Lee, K., Jo, S., and Lee, J. (2022), "The beta-mixture shrinkage prior for sparse covariances with near-minimax posterior convergence rate",
\emph{Journal of Multivariate Analysis}, 192, 105067.
}
\seealso{
sbmspcov
}
\author{
Kyoungjae Lee and Seongil Jo
}
\keyword{covariance}
\keyword{sparse}
