% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.colon.R
\docType{data}
\name{colon}
\alias{colon}
\title{colon dataset}
\format{
data.frame
}
\source{
http://genomics-pubs.princeton.edu/oncology/affydata/.
}
\description{
The colon cancer dataset, which includes gene expression values from 22 colon tumor tissues and 40 non-tumor tissues.
}
\examples{

data("colon")
head(colon)

}
\keyword{data}
