% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.SP500.R
\docType{data}
\name{SP500}
\alias{SP500}
\title{SP500 dataset}
\format{
`list`
}
\source{
State Street Global Advisors.
}
\description{
The S\&P 500 dataset from State Street Global Advisors with the collection period from Jan 2013 to Nov 2023.
}
\examples{

data("SP500")
names(SP500)

}
\keyword{data}
