% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.SP500_idioerr.R
\docType{data}
\name{SP500_idioerr}
\alias{SP500_idioerr}
\title{SP500 dataset}
\format{
`list`
}
\source{
State Street Global Advisors (2022).
}
\description{
The idiosyncratic error of S&P 500 data
}
\examples{

\dontrun{
# SP500_idioerr <- readRDS("SP500_idioerr.rds")
# devtools::use_data(SP500_idioerr, overwrite = TRUE)
data("SP500_idioerr")
names(SP500_idioerr)
}
}
\keyword{data}
