% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tissues.R
\docType{data}
\name{tissues}
\alias{tissues}
\title{tissues dataset}
\format{
`numeric`
}
\source{
http://genomics-pubs.princeton.edu/oncology/affydata/.
}
\description{
The tissues data of colon cancer dataset,
which includes gene expression values from 22 colon tumor tissues and 40 non-tumor tissues.
}
\examples{

\dontrun{
# tissues <- scan("tissues.txt")
# devtools::use_data(tissues, overwrite = TRUE)
data("tissues")
head(tissues)
}
}
\keyword{data}
